+++ app.js
@@ -19,5 +19,6 @@ App({
     var logs = wx.getStorageSync('logs') || []
     logs.unshift(Date.now())
     wx.setStorageSync('logs', logs)
-  }
+  },
+
 })
