+++ pages/collect/collect.json
@@ -1,0 +1,1 @@
+{}
