+++ pages/collect/collect.wxml
@@ -1,0 +1,2 @@
+<!--pages/collect/collect.wxml-->
+<text>pages/collect/collect.wxml</text>
