+++ pages/collect/collect.wxss
@@ -1,0 +1,1 @@
+/* pages/collect/collect.wxss */
