+++ pages/profile/profile.json
@@ -1,1 +1,4 @@
-{}
+{
+    "enablePullDownRefresh": false,
+    "navigationBarTitleText": "个人中心"
+  }
