+++ pages/user_upload/user_upload.wxml
@@ -1,0 +1,2 @@
+<!--pages/user_upload/user_upload.wxml-->
+<text>pages/user_upload/user_upload.wxml</text>
