+++ pages/user_upload/user_upload.wxss
@@ -1,0 +1,1 @@
+/* pages/user_upload/user_upload.wxss */
